Azure connection and pagination added

+++ Tenant Rent Backend/src/main/java/com/springRest/tenantRent/repository/RentPaymentRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RentPaymentRepository extends JpaRepository<RentPayment, Integer> {
+public interface RentPaymentRepository extends JpaRepository<RentPayment, Long> {
     // Custom query methods can be added here if needed
     
 }
